Refactored tests to rspec
