+++ README.md
@@ -7,7 +7,7 @@ Zidox specific integration for Redmine.
 Clone the repository in the `plugins` direcotry, run migrations and restart
 redmine:
 ```
-git clone git@github.com:sdwolf/redmine_resources.git
+git clone git@github.com:ZitecCOM/redmine_resources.git
 rake redmine:plugins:migrate NAME=redmine_resources
 ```
 
@@ -25,13 +25,13 @@ defaults.
 
 Make sure you have the testing gems plugin:
 ```
-git clone git@github.com:sdwolf/redmine_testing_gems.git
+git clone git@github.com:ZitecCOM/redmine_testing_gems.git
 bundle install
 ```
 
 Then run:
 ```
-rake redmine:plugins:test NAME=redmine_resources
+rake redmine:plugins:spec RAILS_ENV=test NAME=redmine_resources
 ```
 
 To view test coverage go to `plugins/redmine_resources/tmp/coverage`
