+++ spec/factories/issue_resource.rb
