+++ spec/factories/resource.rb
@@ -1,6 +1,6 @@
 FactoryGirl.define do
   factory :resource do
-    sequence(:name) {|count| "Sample Name #{ count }" }
     sequence(:code) {|count| "Sample Code #{ count }" }
+    sequence(:name) {|count| "Sample Name #{ count }" }
   end
 end
