+++ spec/factories/resource_setting.rb
@@ -1,6 +1,6 @@
 FactoryGirl.define do
   factory :resource_setting do
-    setting 1
     setting_object_type 'Tracker'
+    setting 1
   end
 end
