+++ spec/spec_helper.rb
@@ -1,0 +1,3 @@
+# Using the default helper from redmine_testing_gems
+helper = File.expand_path '../../../redmine_testing_gems/spec_helper', __FILE__
+require helper
