--- test/factories/division.rb
@@ -1,5 +1,0 @@
-FactoryGirl.define do
-  factory :division do
-    sequence(:name) {|count| "Sample Name #{ count }" }
-  end
-end
