--- test/factories/issue_status.rb
@@ -1,6 +1,0 @@
-FactoryGirl.define do
-  factory :issue_status do
-    sequence(:name) {|count| "Sample Name #{ count }" }
-    is_closed false
-  end
-end
