--- test/factories/member.rb
@@ -1,4 +1,0 @@
-FactoryGirl.define do
-  factory :member do
-  end
-end
