--- test/factories/member_role.rb
@@ -1,4 +1,0 @@
-FactoryGirl.define do
-  factory :member_role do
-  end
-end
