--- test/factories/workflow_permission.rb
@@ -1,8 +1,0 @@
-FactoryGirl.define do
-  factory :workflow_permission do
-    type "WorkflowPermission"
-    assignee false
-    author false
-    new_status_id 0
-  end
-end
