--- test/functional/issues_controller_test.rb
@@ -1,17 +1,0 @@
-require File.expand_path('../../test_helper', __FILE__)
-
-class IssuesControllerTest < ActionController::TestCase
-  def create_test_setup
-    @hours = 2
-    login_as_admin
-    create_base_setup_with_settings
-    create_issue_with_initial_estimation_of @hours
-  end
-
-  test 'GET #show' do
-    create_test_setup
-    get :show, id: @issue.id
-    assert_response :success
-    assert_select 'span', @custom_field.name
-  end
-end
